+++ R/api.R
@@ -27,7 +27,8 @@
 #' \donttest{
 #' # fetch all data for a given flow and key
 #' bbk_data("BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A")
-#'
+#' # fetch data for multiple keys
+#' bbk_data("BBK01", c("TTA032", "TTA010"))
 #' # specified period (start date-end date) for daily data
 #' bbk_data(
 #'   "BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A",
@@ -48,7 +49,7 @@ bbk_data <- function(flow,
                      last_n = NULL) {
   stopifnot(
     is_string(flow), nchar(flow) %in% 5:8,
-    is_string_or_null(key),
+    is_character_or_null(key),
     is_string_or_null(start_period),
     is_string_or_null(end_period),
     is_count_or_null(first_n),
@@ -60,6 +61,7 @@ bbk_data <- function(flow,
     resource <- sprintf("data/%s", flow)
   } else {
     key <- toupper(key)
+    key <- paste(key, collapse = "+")
     resource <- sprintf("data/%s/%s", flow, key)
   }
   body <- make_request(