+++ R/assertions.R
@@ -2,6 +2,10 @@ is_character <- function(x) {
   is.character(x) && !anyNA(x) && length(x) > 0L
 }
 
+is_character_or_null <- function(x) {
+  is.null(x) || is_character(x)
+}
+
 is_string <- function(x) {
   is.character(x) && length(x) == 1L && !is.na(x)
 }