+++ NEWS.md
@@ -3,6 +3,7 @@
 * Dynamically parse column names in `bb_data()`
 * Support for `/data/flowRef` endpoint
 * Prefix in function name has changed from `bb_` to `bbk_`
+* Allow multiple keys in `bbk_data()`
 
 # bbk 0.2.0